+++ example_with_integration_test/lib/models/todo_model.g.dart
@@ -1,0 +1,65 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'todo_model.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+Todo _$TodoFromJson(Map<String, dynamic> json) {
+  return Todo()
+    ..id = json['id'] as String
+    ..created = json['created'] == null
+        ? null
+        : DateTime.parse(json['created'] as String)
+    ..updated = json['updated'] == null
+        ? null
+        : DateTime.parse(json['updated'] as String)
+    ..action = json['action'] as String
+    ..status = _$enumDecodeNullable(_$TodoStatusEnumMap, json['status']);
+}
+
+Map<String, dynamic> _$TodoToJson(Todo instance) => <String, dynamic>{
+      'id': instance.id,
+      'created': instance.created?.toIso8601String(),
+      'updated': instance.updated?.toIso8601String(),
+      'action': instance.action,
+      'status': _$TodoStatusEnumMap[instance.status],
+    };
+
+T _$enumDecode<T>(
+  Map<T, dynamic> enumValues,
+  dynamic source, {
+  T unknownValue,
+}) {
+  if (source == null) {
+    throw ArgumentError('A value must be provided. Supported values: '
+        '${enumValues.values.join(', ')}');
+  }
+
+  final value = enumValues.entries
+      .singleWhere((e) => e.value == source, orElse: () => null)
+      ?.key;
+
+  if (value == null && unknownValue == null) {
+    throw ArgumentError('`$source` is not one of the supported values: '
+        '${enumValues.values.join(', ')}');
+  }
+  return value ?? unknownValue;
+}
+
+T _$enumDecodeNullable<T>(
+  Map<T, dynamic> enumValues,
+  dynamic source, {
+  T unknownValue,
+}) {
+  if (source == null) {
+    return null;
+  }
+  return _$enumDecode<T>(enumValues, source, unknownValue: unknownValue);
+}
+
+const _$TodoStatusEnumMap = {
+  TodoStatus.pending: 'pending',
+  TodoStatus.complete: 'complete',
+};